// apps/git-page/src/app/enums.ts
export enum MainRouteName {
  Home = 'home',
  DemoCheckbox = 'demo-checkbox',
  DemoRadio = 'demo-radio',
  API = 'api',
  Install = 'install'
}

export enum InstallRouteName {
  Lib = 'lib',
  ScssCustomize = 'scss-customize'
}

export enum DemoRadioRouteName {
  All = 'all',
  Basic = 'basic',
  Animations = 'animations',
  Color = 'color',
  ColorOutlineIcon = 'color-outline-icon',
  ColorIcon = 'color-icon',
  Plain = 'plain',
  Switch = 'switch',
}

export enum DemoCheckboxesRouteName {
  Basic = 'basic',
  Switch = 'switch',
  Colors = 'colors',

  MixedColor = 'mixed-color',
  MixedOutline = 'mixed-outline',
  MixedSquareFill = 'mixed-square-fill',
  MixedSquareThick = 'mixed-square-thick',
  MixedSquareThickOutline = 'mixed-square-thick-outline',
  MixedCurve = 'mixed-curve',
  MixedCurveOutline = 'mixed-curve-outline',
  MixedCurveFill = 'mixed-curve-fill',
  MixedCurveThick = 'mixed-curve-thick',
  MixedRound = 'mixed-round',
  MixedRoundFill = 'mixed-round-fill',
  MixedRoundThick = 'mixed-round-thick',
  MixedSwitch = 'mixed-switch',

  FontIcons = 'font-icons',
  Svg = 'svg',
  Image = 'image',

  Animations = 'animations',
  AnimationSmooth = 'animation-smooth',
  AnimationJelly = 'animation-jelly',
  AnimationTada = 'animation-tada',
  AnimationRotate = 'animation-rotate',
  AnimationPulse = 'animation-pulse',

  Toggle = 'toggle',
  Scalability = 'scalability',
  States = 'states',
  All = 'all',

  Plain = 'plain',
  Disable = 'disable',
  Lock = 'lock',
  Size = 'size',
}

export enum ApiRouteName {
  Checkbox = 'checkbox',
  CheckboxWillChange = 'checkbox-will-change',
  RadioGroup = 'radio-group',
  Radio = 'radio',
  RadioWillChange = 'radio-will-change',
  Interfaces = 'interfaces',
  Enums = 'enums',
  Toggle = 'toggle',

  ToggleWillChange = 'toggle-will-change',
  Hover = 'hover',
  HoverWillChange = 'hover-will-change',
  Indeterminate = 'indeterminate',
  IndeterminateWillChange = 'indeterminate-will-change',
  Icon = 'icon',
  Svg = 'svg',
  Image = 'image'
}